Created 'useIsMobile' hook.
